Move identifier to centroided spectra.
